requirements.txt
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
Automat==25.4.16
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
constantly==23.10.4
cryptography==45.0.4
cssselect==1.3.0
defusedxml==0.7.1
filelock==3.18.0
grpcio==1.73.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
itemadapter==0.11.0
itemloaders==1.3.2
jmespath==1.0.1
lxml==5.4.0
numpy==2.3.0
packaging==25.0
parsel==1.10.0
portalocker==2.10.1
Protego==0.4.0
protobuf==6.31.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyDispatcher==2.0.7
pyOpenSSL==25.1.0
python-dateutil==2.9.0.post0
qdrant-client==1.14.3
queuelib==1.8.0
requests==2.32.4
requests-file==2.1.0
Scrapy==2.13.2
service-identity==24.2.0
six==1.17.0
sniffio==1.3.1
tldextract==5.3.0
Twisted==25.5.0
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
w3lib==2.3.1
zope.interface==7.2
